Add files via upload
This project is a real-time face emotion detection system built with Python. It uses OpenCV and a pre-trained deep learning model to detect human faces through a webcam and classify emotions like happy, sad, angry, surprised, and neutral. The detected emotion is displayed with a corresponding emoji in a simple Tkinter GUI. The system supports multiple face detection and provides a fun, interactive way to understand facial expressions.